requirements.txt
streamlit
pandas
openpyxl
pythonocc-core